TcpIpAcceptor owns intercept logic
So the logic was moved into the IOService and this makes the logic hard
to comprehend. The whole logic has now been moved into the TcpIpAcceptor
so that the logic is easier to understand because it is smeared out
over less places.
